Add descriptions to rsnlinkingCommands, update PM text
Added some basic descriptions for the rsnlinkingCommands that are displayed when the !help command is called. 

Replaced instances of !link and !verify with /link and /verify (!help still displays these commands as !link and !verify, I'm not familiar with how command prefixes work, so I'm unsure how to fix that.) 

Updated PM instructions to refer to slightly more accurate channel names - unsure if emojis are supported in this field- and reference the possibility of creating a ticket in #support when seeking verification.